MLIBZ-2351: Use webview events to capture redirect
