--- src/nativescript/popup/package.json
@@ -1,6 +1,0 @@
-{
-  "name" : "popup",
-  "main" : "popup",
-  "types": "popup.d.ts",
-  "nativescript": {}
-}
