--- src/nativescript/popup/popup.d.ts
@@ -1,6 +1,0 @@
-import { EventEmitter } from 'events';
-
-export declare class Popup extends EventEmitter {
-  open(url: string): Popup;
-  close(): Popup;
-}
